requirements.txt
bidict==0.23.1
black==24.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
flake8==7.1.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fluvio==0.20.0
greenlet==3.2.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==1.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.12.1
pyflakes==3.2.0
pymongo==4.12.0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.32.3
semantic-version==2.10.0
setuptools==79.0.0
setuptools-rust==1.10.1
simple-websocket==1.1.0
toml==0.10.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
